requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-debug-toolbar==3.2
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
whitenoise==5.2.0
